ai_service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
mlxtend>=0.23.1
langchain>=0.1.13
langchain-groq>=0.1.1
langchain-community>=0.0.29
sqlalchemy>=2.0.28